Revert "A separate agent permission management administrator service added."
This reverts commit c1735ce00de8a9d3a19e81082e67fea24e834589.
