+++ Api/Services/Agents/IAgentPermissionManagementService.cs
@@ -5,10 +5,12 @@ using HappyTravel.Edo.Common.Enums;
 
 namespace HappyTravel.Edo.Api.Services.Agents
 {
-    public interface IAgentPermissionManagementService<T>
+    public interface IAgentPermissionManagementService
     {
-        Task<Result<List<InAgencyPermissions>>> SetInAgencyPermissions(int agencyId, int agentId, List<InAgencyPermissions> permissions);
+        Task<Result<List<InAgencyPermissions>>> SetInAgencyPermissions(
+            int agencyId, int agentId, List<InAgencyPermissions> permissions);
 
-        Task<Result<List<InAgencyPermissions>>> SetInAgencyPermissions(int agencyId, int agentId, InAgencyPermissions permissions);
+        Task<Result<List<InAgencyPermissions>>> SetInAgencyPermissions(
+            int agencyId, int agentId, InAgencyPermissions permissions);
     }
 }
